적 움직임 구현 + 세부적인 코드 삽입 요망

+++ AlphaEngine/Components/PlayerComp.h
@@ -15,7 +15,7 @@ public:
 	Data::PlayerData data;
 	bool moveState = true;
 	bool turnTemp = true;
-	float GetMovegauge();
+	int GetMovegauge();
 	PlayerComp(GameObject* _owner);
 	~PlayerComp();
 	void Update() override;
